moved validateParameterValue to helper
